Moved q classes to their own file for better organization
