+++ .pre-commit-config.yaml
@@ -30,7 +30,7 @@ repos:
     rev: "v2.2.6"
     hooks:
       - id: codespell
-        args: ["src/*"]
+        args: ["src/*", "--", "--ignore-word-list", "ket"]
 
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: "v1.9.0"
